restructured integration package
