+++ src/main/java/com/bafomdad/uniquecrops/integration/jei/UCRecipeHandler.java
@@ -1,4 +1,4 @@
-package com.bafomdad.uniquecrops.integration.craftyplants;
+package com.bafomdad.uniquecrops.integration.jei;
 
 import com.bafomdad.uniquecrops.crafting.SeedRecipe;
